split core into web and hooks
